15/01/2024. Se avanza proyecto alumnos

+++ tema-07/proyectos/01-validacion/07/mvc-proyect/views/alumno/main/index.php
@@ -91,7 +91,7 @@
 								<!-- botones de acción -->
 								<td>
 									<!-- botón  eliminar -->
-									<a href="<?= URL ?>alumno/delete/<?= $alumno->id ?>" title="Eliminar">
+									<a href="<?= URL ?>alumno/delete/<?= $alumno->id ?>" title="Eliminar" onclick="return confirm('Confirme la eliminación del alumno')">
 										<i class="bi bi-trash-fill"></i></a>
 
 									<!-- botón  editar -->
